# Conflicts:
#	CMakeLists.txt
#	roboteam_ai/src/conditions/IHaveBall.cpp
#	roboteam_ai/src/conditions/IsInDefenseArea.cpp
#	roboteam_ai/src/conditions/IsInDefenseArea.hpp
#	roboteam_ai/src/conditions/IsInZone.cpp
#	roboteam_ai/src/conditions/IsRobotClosestToBall.cpp
#	roboteam_ai/src/conditions/IsRobotClosestToBall.h
#	roboteam_ai/src/utilities/Field.cpp
#	roboteam_ai/src/utilities/Field.h
#	roboteam_ai/src/utilities/World.cpp
#	roboteam_ai/src/utilities/World.h
#	roboteam_ai/test/ConditionTests/IHaveBallTest.cpp
#	roboteam_ai/test/ConditionTests/IsInDefenseAreaTest.cpp
#	roboteam_ai/test/ConditionTests/IsInZoneTest.cpp

+++ roboteam_ai/src/utilities/Field.cpp
@@ -10,39 +10,40 @@ namespace ai {
 roboteam_msgs::GeometryFieldSize Field::field;
 
 const roboteam_msgs::GeometryFieldSize Field::get_field() {
-  return Field::field;
+    return Field::field;
 }
 
 void Field::set_field(roboteam_msgs::GeometryFieldSize field) {
-  Field::field = field;
+    Field::field = field;
 }
 
 Vector2 Field::get_our_goal_center() {
-  return Vector2(field.field_length / -2, 0);
+    return Vector2(field.field_length/- 2, 0);
 }
 
 Vector2 Field::get_their_goal_center() {
-  return Vector2(field.field_length / 2, 0);
+    return Vector2(field.field_length/2, 0);
 }
 
-bool Field::pointIsInDefenceArea(Vector2 point, bool isOurDefenceArea, double margin) {
-  auto penaltyLine = isOurDefenceArea ? field.left_penalty_line : field.right_penalty_line;
-  double yTopBound;
-  double yBottomBound;
-  double xBound = penaltyLine.begin.x;
-  if (penaltyLine.begin.y < penaltyLine.end.y) {
-    yBottomBound = penaltyLine.begin.y;
-    yTopBound = penaltyLine.end.y;
-  } else {
-    yBottomBound = penaltyLine.end.y;
-    yTopBound = penaltyLine.begin.y;
-  }
-  bool yIsWithinDefenceArea = point.y < (yTopBound + margin) && point.y > (yBottomBound - margin);
-  bool xIsWithinOurDefenceArea = point.x < (xBound + margin);
-  bool xIsWithinTheirDefenceArea = point.x > (xBound - margin);
+bool Field::pointIsInDefenceArea(Vector2 point, bool isOurDefenceArea, float margin) {
+    auto penaltyLine = isOurDefenceArea ? field.left_penalty_line : field.right_penalty_line;
+    double yTopBound;
+    double yBottomBound;
+    double xBound = penaltyLine.begin.x;
+    if (penaltyLine.begin.y < penaltyLine.end.y) {
+        yBottomBound = penaltyLine.begin.y;
+        yTopBound = penaltyLine.end.y;
+    }
+    else {
+        yBottomBound = penaltyLine.end.y;
+        yTopBound = penaltyLine.begin.y;
+    }
+    bool yIsWithinDefenceArea = point.y<(yTopBound + margin) && point.y>(yBottomBound - margin);
+    bool xIsWithinOurDefenceArea = point.x < (xBound + margin);
+    bool xIsWithinTheirDefenceArea = point.x > (xBound - margin);
 
-  return yIsWithinDefenceArea && ((isOurDefenceArea && xIsWithinOurDefenceArea)
-  || (!isOurDefenceArea && xIsWithinTheirDefenceArea));
+    return yIsWithinDefenceArea && ((isOurDefenceArea && xIsWithinOurDefenceArea)
+            || (! isOurDefenceArea && xIsWithinTheirDefenceArea));
 }
 
 } // ai
