+++ roboteam_ai/src/utilities/Field.h
@@ -15,15 +15,19 @@ namespace rtt {
 namespace ai {
 
 class Field {
- private:
-  static roboteam_msgs::GeometryFieldSize field;
+    private:
+        static roboteam_msgs::GeometryFieldSize field;
 
- public:
-  static const roboteam_msgs::GeometryFieldSize get_field();
-  static void set_field(roboteam_msgs::GeometryFieldSize field);
-  static Vector2 get_our_goal_center();
-  static Vector2 get_their_goal_center();
-  static bool pointIsInDefenceArea(Vector2 point, bool isOurDefenceArea = true, double margin = 0.0);
+    public:
+        static const roboteam_msgs::GeometryFieldSize get_field();
+
+        static void set_field(roboteam_msgs::GeometryFieldSize field);
+
+        static Vector2 get_our_goal_center();
+
+        static Vector2 get_their_goal_center();
+
+        static bool pointIsInDefenceArea(Vector2 point, bool isOurDefenceArea = true, float margin = 0.0);
 };
 
 } // ai
